Create sensor entry with dispenser name

+++ dashboard/api.py
@@ -9,6 +9,7 @@ from dashboard.models import SensorData
 import paho.mqtt.client as mqtt
 import threading
 
+
 class mqttThread(threading.Thread):
     def __init__(self, threadID, client):
         threading.Thread.__init__(self)
@@ -21,6 +22,8 @@ class mqttThread(threading.Thread):
         self.client.connect("stevenhuang.ca", 12001, 60)
         self.client.publish("/test", "csa-site client mqtt connected")
         self.pulls = 0
+        self.dispenser = {}
+
 
     def run(self):
         print("MQTT Thread starting...")
@@ -34,26 +37,34 @@ class mqttThread(threading.Thread):
     def on_connect(self, client, userdata, flags, rc):
         print("Connected with result code "+str(rc))
         #self.client.subscribe("$SYS/#")
-        # Subscribe to everything
-        self.client.subscribe("#")
+        self.client.subscribe("dispenser")
+        self.client.subscribe("connCheck")
+        self.client.subscribe("action")
 
     # The callback for when a PUBLISH message is received from the server.
     def on_message(self, client, userdata, msg):
-        f=""
         topics = msg.topic.split("/")
         try:
             # msg from a module
-            if topics[0] == "dispensors" and len(topics) == 2:
-                if msg.payload == "pull":
-                    print("Detected pull from %s" % topics[1])
-                    print(topics)
-                    ff = SensorData.create(topics[1])
-                    ff.save()
-                    self.pulls += 1
+            if topics[0] == "dispenser":
+                if len(topics) == 2:
+                    if msg.payload == "pull":
+                        print("Detected pull from %s" % topics[1])
+                        newEntry = SensorData.create(name=topics[1])
+                        newEntry.save()
+                        self.pulls += 1
+                    elif msg.payload == "alive":
+                        print("Alive: %s" % topics[1])
+                        #getattr(self.dispenser.
+                    elif msg.payload == "dead":
+                        print("Dead: %s" % topics[1])
+                elif len(topics) == 3:
+                    if topics[2] == "status":
+                        print("Recieved status update from" % topics[1])
             # TODO: authorize this
             elif topics[0] == "action":
                 if topics[1] == "delete":
-                    if topics[2] == "entries":
+                    if topics[2] == "entries" and msg.payload == "yesrly":
                         print("Deleting all entries")
                         SensorData.objects.all().delete()
         except IndexError:
@@ -133,7 +144,7 @@ class StatisticsResource(Resource):
         f += 1
         setattr(setting, 'test_value', f)
         setattr(setting, 'random', randint(1, 1000))
-        setattr(setting, 'cost_current', len(SensorData.objects.all()) * 0.01)
+        setattr(setting, 'cost_current', SensorData.objects.count() * 0.01)
         setattr(setting, 'cost_projected_daily', randint(3600, 3750) + f)
         return setting
